apps/task/serializers.py
from rest_framework import serializers

from .models import Role, Task


class RoleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Role
        fields = ['id', 'name', 'description']


class TaskSerializer(serializers.ModelSerializer):
    created_by = serializers.StringRelatedField()

    class Meta:
        model = Task
        fields = "__all__"
