requirements.txt
django==3.2
djangorestframework==3.14
celery==5.3.4
python-decouple==3.8
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
loguru==0.7.1
drf-yasg==1.21.7
redis==5.0.0